Added timing code
Now win32 has timing methods to retrieve a delta change in time since
the last call. The change is in milliseconds.

You must call SetupTimer() before you call GetDeltaTime() in order to
establish the timer frequency.

+++ src/ISystem.h
@@ -26,7 +26,7 @@ public:
 	 * \returns   void 
 	 * \exception  
 	 */
-	virtual void Update(const float delta) = 0;
+	virtual void Update(const double delta) = 0;
 
 	/**
 	 * \brief Retrieve the component that belongs to the given entity ID
